Moves tests outside of build sources.
