+++ packages/utils/common/src/__tests__/isEmpty.test.ts
@@ -1,4 +1,4 @@
-import {isEmpty} from '../isEmpty';
+import {isEmpty} from '../ts/isEmpty';
 
 test('array', () => {
     expect(isEmpty([])).toBeTruthy();
